+++ lib/src/sprite.dart
@@ -13,7 +13,7 @@ class Sprite {
 
   Sprite({this.image, this.imageRect, this.transform});
 
-  Future<Sprite> load() async {
+  Future<Sprite> loadImage() async {
     image = await ImageAssets.instance.load(imageRect.image);
     if (imageRect.rect == null) {
       imageRect.rect = IntRect(0, 0, image.width, image.height);
@@ -23,7 +23,7 @@ class Sprite {
 
   static Future<Sprite> loadJson(String name) async {
     final jsonSprite = await JsonAssets.instance.load(name);
-    return await Sprite.fromJson(jsonSprite).load();
+    return await Sprite.fromJson(jsonSprite).loadImage();
   }
 
   factory Sprite.fromJson(Map<String, dynamic> json, {String image}) {